+++ internal/app/init/main.go
@@ -26,7 +26,7 @@ import (
 	"github.com/siderolabs/talos/internal/pkg/mount/switchroot"
 	"github.com/siderolabs/talos/internal/pkg/rng"
 	"github.com/siderolabs/talos/internal/pkg/secureboot"
-	"github.com/siderolabs/talos/internal/pkg/tpm2"
+	"github.com/siderolabs/talos/internal/pkg/secureboot/tpm2"
 	"github.com/siderolabs/talos/pkg/machinery/constants"
 	"github.com/siderolabs/talos/pkg/machinery/extensions"
 	"github.com/siderolabs/talos/pkg/version"
@@ -61,8 +61,8 @@ func run() (err error) {
 	}
 
 	// extend PCR 11 with enter-initrd
-	if err = tpm2.PCRExtent(constants.UKIMeasuredPCR, []byte(secureboot.EnterInitrd)); err != nil {
-		return fmt.Errorf("failed to extend PCR %d with enter-initrd: %v", constants.UKIMeasuredPCR, err)
+	if err = tpm2.PCRExtent(secureboot.UKIPCR, []byte(secureboot.EnterInitrd)); err != nil {
+		return fmt.Errorf("failed to extend PCR %d with enter-initrd: %v", secureboot.UKIPCR, err)
 	}
 
 	log.Printf("booting Talos %s", version.Tag)
