--- internal/pkg/secureboot/measure/internal/pcr/policy.go
@@ -1,34 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-package pcr
-
-import (
-	"crypto/sha256"
-
-	"github.com/google/go-tpm/tpm2"
-)
-
-// CalculatePolicy calculates the policy hash for a given PCR value and PCR selection.
-func CalculatePolicy(pcrValue []byte, pcrSelection tpm2.TPMLPCRSelection) ([]byte, error) {
-	calculator, err := tpm2.NewPolicyCalculator(tpm2.TPMAlgSHA256)
-	if err != nil {
-		return nil, err
-	}
-
-	pcrHash := sha256.Sum256(pcrValue)
-
-	policy := tpm2.PolicyPCR{
-		PcrDigest: tpm2.TPM2BDigest{
-			Buffer: pcrHash[:],
-		},
-		Pcrs: pcrSelection,
-	}
-
-	if err := policy.Update(calculator); err != nil {
-		return nil, err
-	}
-
-	return calculator.Hash().Digest, nil
-}
