+++ internal/pkg/secureboot/secureboot.go
@@ -78,5 +78,10 @@ func OrderedPhases() []PhaseInfo {
 	}
 }
 
-// UKIPCR is the PCR number where sections except `.pcrsig` are measured.
-const UKIPCR = 11
+const (
+	// UKIPCR is the PCR number where sections except `.pcrsig` are measured.
+	UKIPCR = 11
+	// SecureBootStatePCR is the PCR number where the secure boot state and the signature are measured.
+	// PCR 7 changes when UEFI SecureBoot mode is enabled/disabled, or firmware certificates (PK, KEK, db, dbx, …) are updated.
+	SecureBootStatePCR = 7
+)
